bugfixes and additions
Image viewer now shows images in order and displays the image name over each of the images. Added a button to take single images outside of events. The default save folder is the dump folder to avoid accidentally overwriting saved images -- the sym link is reset to the appropriate Images folder only when an event runs properly.